[feature] Project refactor (#66)
* [feature] Project refactor model

* [feature] Project refactor model sql

* [feature] Move deprecated packages in model

* [feature] Project detail info, expenses, signings and project leaders

* [feature] Project customer refactor

* [feature] Project member refactor

* [feature] Project materials

* [feature] Project families and start clear code

* [feature] Clear code

* [feature] User clear code

+++ gestepm-forum-model-api/src/main/java/com/epm/gestepm/forum/model/api/service/TopicService.java
@@ -1,12 +1,11 @@
 package com.epm.gestepm.forum.model.api.service;
 
+import com.epm.gestepm.forum.model.api.dto.Topic;
+import org.springframework.web.multipart.MultipartFile;
+
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
 
-import org.springframework.web.multipart.MultipartFile;
-
-import com.epm.gestepm.forum.model.api.dto.Topic;
-
 public interface TopicService {
-	CompletableFuture<Topic> create(String topicTitle, String topicContent, Long forumId, String userIp, String username, List<MultipartFile> files);
+	CompletableFuture<Topic> create(String topicTitle, String topicContent, Integer forumId, String userIp, String username, List<MultipartFile> files);
 }
